+++ pallets/ema-oracle/src/tests/mod.rs
@@ -18,7 +18,7 @@
 mod add_and_remove_oracle;
 mod invariants;
 mod mock;
-mod update_bitfrost_oracle;
+mod update_bifrost_oracle;
 
 use super::*;
 use crate::OraclePeriod::*;
